requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
exceptiongroup==1.1.1
fake-useragent==1.1.1
googlesearch-python==1.1.0
h11==0.14.0
html5lib==1.1
idna==2.10
lxml==4.9.2
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
PySimpleGUI==4.60.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.25.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
trio==0.22.0
trio-websocket==0.10.0
urllib3==1.26.15
webencodings==0.5.1
wsproto==1.2.0
